requirements.txt
toml>=0.10.2
requests>=2.31.0
python-dotenv>=1.0.0